requirements.txt
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
gevent==21.1.2
greenlet==1.0.0
idna==2.10
pycparser==2.20
pyserial==3.5
python-engineio==3.14.0
python-engineio-client==1.1
python-socketio==5.0.4
python-socketio-client==1.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
zope.event==4.5.0
zope.interface==5.2.0
